+++ servlib/bundling/BundleProtocol.cc
@@ -135,7 +135,7 @@ BundleProtocol::format_headers(const Bundle* bundle, u_char* buf, size_t len)
     add_to_dict(bundle->replyto_,   &dict, &dictionary_len);
 
     (void)log; // in case NDEBUG is defined
-    log_debug(log, "generated dictionary length %u", (u_int)dictionary_len);
+    log_debug(log, "generated dictionary length %zu", dictionary_len);
 
     primary_len += SDNV::encoding_len(dictionary_len);
     primary_len += dictionary_len;
@@ -165,9 +165,9 @@ BundleProtocol::format_headers(const Bundle* bundle, u_char* buf, size_t len)
 	return -1;
     }
 
-    log_debug(log, "preamble length %u, primary length %u",
-	      (u_int)(sizeof(PrimaryHeader1) + SDNV::encoding_len(primary_len)),
-	      (u_int)primary_len);
+    log_debug(log, "preamble length %zu, primary length %zu",
+	      (sizeof(PrimaryHeader1) + SDNV::encoding_len(primary_len)),
+	      primary_len);
 
     /*
      * Ok, stuff in the preamble and the total header length.
@@ -280,7 +280,7 @@ BundleProtocol::format_headers(const Bundle* bundle, u_char* buf, size_t len)
     len -= encoding_len;
 
     // return the total buffer length consumed
-    log_debug(log, "encoding done -- total length %u", (u_int)(orig_len - len));
+    log_debug(log, "encoding done -- total length %zu", (orig_len - len));
     return orig_len - len;
 }
 
@@ -298,8 +298,8 @@ BundleProtocol::parse_headers(Bundle* bundle, u_char* buf, size_t len)
     PrimaryHeader1* primary1;
     if (len < sizeof(PrimaryHeader1)) {
  tooshort1:
-	log_debug(log, "buffer too short (parsed %u/%u)",
-		  (u_int)(origlen - len), (u_int)origlen);
+	log_debug(log, "buffer too short (parsed %zu/%zu)",
+		  (origlen - len), origlen);
 	return -1;
     }
 
@@ -349,7 +349,7 @@ BundleProtocol::parse_headers(Bundle* bundle, u_char* buf, size_t len)
     if (len < sizeof(PrimaryHeader2)) {
  tooshort2:
 	log_err(log, "primary header advertised incorrect length: "
-		"advertised %u, total buffer %d", primary_len, (u_int)len);
+		"advertised %u, total buffer %zu", primary_len, len);
 	return -1;
     }
 
@@ -485,8 +485,8 @@ BundleProtocol::parse_headers(Bundle* bundle, u_char* buf, size_t len)
 	switch (preamble->type) {
 	case HEADER_PAYLOAD: {
 	    bundle->payload_.set_length(header_len);
-	    log_debug(log, "parsed payload length %u",
-		      (u_int)bundle->payload_.length());
+	    log_debug(log, "parsed payload length %zu",
+		      bundle->payload_.length());
 
 	    // note that we don't actually snarf in the payload here
 	    // since it's handled by the caller, and since the payload
